Add Event Button Fix
Now the Add Event Button on the homepage is fixed.

When clicked, it should send the user to an add event page that allows the user to add the Name of event, the description of event, the date of event, and the start/end time of the event.

The second "add event" button is not done yet. For now, this button will send to the weekly view layout.

The home button will return the user back into the home screen layout.

Next time:
-Second Add button should gather all information from user and put it in event. Displays it in weekly view. Figure out how to store it.